Move internal packages to pkg, to expose Build* methods
This commit moves the internal packages out to pkg, to make them
importable.

Right now only panels are importable https://pkg.go.dev/github.com/perses/community-dashboards@v0.0.0-20250415165833-742a315148e7
But there are usecases downstream that could benefit from importing the
`Build*` methods or `Exec` methods.

Let me know what you think!
