+++ pkg/dashboards/node_exporter/node_exporter_cluster_use_method.go
@@ -1,9 +1,9 @@
 package nodeexporter
 
 import (
-	"github.com/perses/community-dashboards/internal/dashboards"
-	"github.com/perses/community-dashboards/internal/promql"
+	"github.com/perses/community-dashboards/pkg/dashboards"
 	panels "github.com/perses/community-dashboards/pkg/panels/node_exporter"
+	"github.com/perses/community-dashboards/pkg/promql"
 	"github.com/perses/perses/go-sdk/dashboard"
 	panelgroup "github.com/perses/perses/go-sdk/panel-group"
 	labelValuesVar "github.com/perses/perses/go-sdk/prometheus/variable/label-values"
